added rebuild logic, and multi-system configuration. Also added the local.nix file to select the system configuration.

+++ rebuild.sh
@@ -1,9 +1,18 @@
 #!/run/current-system/sw/bin/bash
 
 sudo cp /etc/nixos/configuration.nix /etc/nixos/configuration.old
-sudo rm -r /etc/scripts/
+if test -d /etc/scripts; then
+  echo "Removing Scripts Folder"
+  sudo rm -r /etc/scripts/
+fi
 sudo mkdir /etc/scripts/
 sudo cp -r ./scripts/ /etc/
 sudo cp ./configuration.nix /etc/nixos/
 
-sudo nixos-rebuild switch
+if ! test -f /etc/nixos/local.nix; then
+  echo Creating Local Config
+  sudo cp ./local.nix /etc/nixos/
+  echo Please edit local nix file and rerun.
+else
+  sudo nixos-rebuild switch
+fi
